+++ src/main/java/bo/edu/ucbcba/prestamix/group4/controller/StoreController.java
@@ -59,10 +59,17 @@ public class StoreController
 
     public void delete(int id)
     {
-        EntityManager entityManager = PrestamixEntityManager.createEntityManager();
-        entityManager.getTransaction().begin();
-        entityManager.remove(entityManager.find(Store.class,id));
-        entityManager.getTransaction().commit();
-        entityManager.close();
+        if(id!=0)
+        {
+            EntityManager entityManager = PrestamixEntityManager.createEntityManager();
+            entityManager.getTransaction().begin();
+            entityManager.remove(entityManager.find(Store.class, id));
+            entityManager.getTransaction().commit();
+            entityManager.close();
+        }
+        else
+        {
+            throw new ValidationException("Error");
+        }
     }
 }
